Debriefing: moved trip checks into utils and added new BN models
